Catch target-host version mismatch
Attach the version information in a symbol that's generated by the host.
Then, expect the same symbol exists in the embedded target. If the
linker can't resolve this, then the host and target should be
incompatible.

I'm overloading the __imxrt_family symbol for this purpose. Good luck
trying to name this from Rust or C!